refactor: 예외를 Global Exception으로 처리
