--- src/main/java/lotto/exception/LottoException.java
@@ -1,12 +1,0 @@
-package lotto.exception;
-
-import lotto.exception.common.BaseException;
-import lotto.exception.message.LottoExceptionMessage;
-
-public class LottoException extends BaseException {
-
-    public LottoException(LottoExceptionMessage message) {
-        super(message);
-    }
-
-}
