--- src/main/java/lotto/exception/MoneyException.java
@@ -1,12 +1,0 @@
-package lotto.exception;
-
-import lotto.exception.common.BaseException;
-import lotto.exception.message.LottoMoneyExceptionMessage;
-
-public class MoneyException extends BaseException {
-
-    public MoneyException(LottoMoneyExceptionMessage message) {
-        super(message);
-    }
-
-}
